+++ apps/licences/.env.development
@@ -1,48 +1,9 @@
 AUTH_SECRET=rCCQNISaAPJOMUFzaOFLFctXq5+vR1IIYcFzs3mX+vM=
 
-OIDC_ISSUER_URL=http://localhost:3000/api/auth/mocks/oidc
+OIDC_ISSUER_URL=http://localhost:3030/api/auth/mocks/oidc
 OIDC_CLIENT_ID=01J3FD4RXC30Y3MP79SDEXEAR4
 OIDC_CLIENT_SECRET=au4roCA8P8jb04IRqrCQi1tmB38unX9w
 OIDC_CLIENT_REDIRECT_URI=http://localhost:3000/api/auth/callback/fims
-OIDC_MOCK_FORCED_ENABLE=true
-
-OIDC_MOCK_JWT_PRIVATE_KEY="-----BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQD+apOISVcduE/d
-rjYV8ccJJ/CZSfmuyms+Qaw0PqTdjOmefe2dAb0m9olYMgwkg/yhjA1Q5RsWYpqO
-0p9l8lGRQAx2HnJ2vDrDa1Rb7n4OO87D50U2xqzBZWJbTF6uHKr+AT+jkoUp3/3m
-jk/p5OtrE1Wa38IXxApag23geQtp53wUEngDtoUUUUqpmDKh1Owixmheh7asmDqq
-Roy5ANkcn4uiTPwVUFoCOiHOkNpuR1l3JcMUHSCBbF4QGqRDb/hsoAFAtrXKPw0m
-sFg2BMpmuCiv2ojB2tYusViPWk27stJMDnny0grcOyfFw1GUA7rSH5HKjEOcHnks
-990crOZlAgMBAAECggEAEMzpZqnPLKl/7qfjLvJLHjXqu7iomgwIZnFQRZfwa6fz
-QnHW59QWUVUc7zvI1P15SJK2y2WBsyn6Ql2yxHjqvYCBAijTDrex2rdC+Boo7Vgy
-OHoUdkBK5q+C/fvuUamXehQLh4n0rLX7roatzMh+6lB5yfMmPksGEwGUVysvH4RV
-1smVsSDhACHMyha20H+xd4f+gfPFaiwonwyHHUxOgJAdKEBzusL2M8Jro6QQs3jt
-8Yl6iYqynoOvX3YXjUIKbA4spuK+oaVRBXybRTJoqRXxevcNdkphMy7hz9Ff6/+N
-0JzVGY88YN2iuPgPiFFktOW4aebfHmz9W5aM4O3FmQKBgQD/xLByJ8yKxarmgZyS
-9WsQYDi0/gv7vtxBR1pcfr9jbMOwckMrkllMcEnErjAnTsI4VtWuE6WuzTJW88Wc
-nmEkH0SowCTPO3KyQimeI+OuC0u7f6cwW41a6lQ59xhb+zGUFNhQraJffs9UKIDk
-OZ9TqNdxoNIKhZjKcwCxvm+Q+QKBgQD+pZLTUJqpGqbANH5bnrgXKoCvSKDMKQu1
-wxjHuqMxSLfsJJm8NZxq8ewj6/iCw2Gc5/lb8KTdPyUFMb9mtnVQ9jQoxuEqMOzh
-H29rirea06z8/reOkEf7F+AwObC1jm7jFJVe0zU0MeYAkIKQWW9yyE3y2C7IdSVh
-FVU9WhcHzQKBgARb+NnpPb47wj7thJqoniuISE8laVTgGubt7WdYLbDnIKq4N9bd
-RrJb05YLlfyFhh/psvilkxRYoPwEaaDSvjscYVIy/C2gCMNrB0SPvgiwXpkSgun7
-FXywM8FNKdgffufUNx+g5fyUr7nyesfpB4MohIuGPy44urLwpNlisooZAoGBAID2
-DWOjQJA/hQ92BkAUI5rn1X/vwi9i4iFNCrGr3ZXKBAs9hfYBGJQ4rKSMjrB6jmbO
-PETJnFTH89GrnglAgi1cYpH8FnAtkul36A+Iku+AH+2IeiFjBHEvaEREEy//UD0E
-YZlCUYF7e8XxaY39HSn+Myu3PAoc+a8vBJork9pJAoGBALzELD+y4IC2UdJv6Lcu
-rzMwsi8HcjjBZoV4QO6GVb7NOsXkCAvEGA0r6dEjBUAWNVQih5UOwb2VcswvsSBq
-HFewaouFUFB167Kv+tNAHZh8LFWkhwPbvwoyrWq2aEoFovRiW/nKfu/ep34Iw8Rf
-QCSdgIH2dSpyvvYaYYgfcJTM
------END PRIVATE KEY-----"
-OIDC_MOCK_JWT_PUBLIC_KEY="-----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA/mqTiElXHbhP3a42FfHH
-CSfwmUn5rsprPkGsND6k3Yzpnn3tnQG9JvaJWDIMJIP8oYwNUOUbFmKajtKfZfJR
-kUAMdh5ydrw6w2tUW+5+DjvOw+dFNsaswWViW0xerhyq/gE/o5KFKd/95o5P6eTr
-axNVmt/CF8QKWoNt4HkLaed8FBJ4A7aFFFFKqZgyodTsIsZoXoe2rJg6qkaMuQDZ
-HJ+Lokz8FVBaAjohzpDabkdZdyXDFB0ggWxeEBqkQ2/4bKABQLa1yj8NJrBYNgTK
-Zrgor9qIwdrWLrFYj1pNu7LSTA558tIK3DsnxcNRlAO60h+RyoxDnB55LPfdHKzm
-ZQIDAQAB
------END PUBLIC KEY-----"
 
 NEXT_PUBLIC_IS_MSW_ENABLED=true
