--- apps/licences/src/app/api/auth/mocks/mocks-util.ts
@@ -1,7 +1,0 @@
-import { NextResponse } from "next/server";
-
-export const getMockForbiddenResponse = () =>
-  NextResponse.json(
-    { message: "API available only in development or test environment." },
-    { status: 403 },
-  );
