+++ apps/op-mock/README.md
@@ -1,0 +1,5 @@
+## OpenID Connect mock provider
+
+This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app), used in this monorepo as a mock oidc provider.
+
+**Note:** _This project is for development and test purposes only._
