+++ apps/op-mock/next.config.mjs
@@ -1,0 +1,11 @@
+/** @type {import('next').NextConfig} */
+
+const nextConfig = {
+  reactStrictMode: true,
+  output: "standalone",
+  experimental: {
+    instrumentationHook: true,
+  },
+};
+
+export default nextConfig;
