+++ apps/op-mock/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "io-ipatente-op-mock",
+  "author": {
+    "name": "PagoPA"
+  },
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "format": "prettier --write .",
+    "format:check": "prettier --check .",
+    "lint": "eslint --fix src",
+    "lint:check": "eslint src",
+    "dev": "next dev -p 3030",
+    "build": "next build",
+    "build:pre-package": "cp -r .next/static .next/standalone/apps/op-mock/.next && cp -r public .next/standalone/apps/op-mock",
+    "build:package": "yarn build:pre-package && (cd .next/standalone/apps/op-mock && zip -r -q ../../../../${npm_package_name}-${npm_package_version}.zip . -x package.json)",
+    "start": "next start -p 3030"
+  },
+  "dependencies": {
+    "jsonwebtoken": "^9.0.2",
+    "next": "^14.2.15",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
+  },
+  "devDependencies": {
+    "@pagopa/eslint-config": "^4.0.1",
+    "@tanstack/eslint-plugin-query": "^5.57.1",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.5",
+    "prettier": "^3.3.3",
+    "typescript": "^5"
+  }
+}
