+++ apps/op-mock/src/app/api/auth/mocks/oidc/authorize/route.ts
@@ -1,0 +1,15 @@
+import type { NextRequest } from "next/server";
+
+import { NextResponse } from "next/server";
+
+export async function GET(request: NextRequest) {
+  // Get query params
+  const redirect_uri = request.nextUrl.searchParams.get("redirect_uri") ?? "";
+  const state = request.nextUrl.searchParams.get("state") ?? "";
+
+  // Simulate a mock authorization code
+  const code = "mock-auth-code";
+
+  // Redirect to client with auth code and state params
+  return NextResponse.redirect(`${redirect_uri}?code=${code}&state=${state}`);
+}
