+++ apps/op-mock/src/app/api/auth/mocks/oidc/route.ts
@@ -1,0 +1,8 @@
+import { NextResponse } from "next/server";
+
+// GET without params are statics, this will force the route to be dynamic
+export const dynamic = "force-dynamic";
+
+export async function GET() {
+  return NextResponse.json({}, { status: 200 });
+}
