+++ apps/op-mock/src/app/api/info/route.ts
@@ -1,0 +1,16 @@
+import { NextResponse } from "next/server";
+
+import packageJson from "../../../../package.json";
+
+/**
+ * `api/info` route handler
+ * @returns project _name_ and _version_
+ */
+export async function GET() {
+  const response = {
+    name: packageJson.name,
+    version: packageJson.version,
+  };
+
+  return NextResponse.json(response, { status: 200 });
+}
