+++ apps/payments/.env.test
@@ -1,6 +1,5 @@
-OIDC_ISSUER_URL=http://localhost:3000/api/auth/mocks/oidc
+OIDC_ISSUER_URL=http://localhost:3030/api/auth/mocks/oidc
 OIDC_CLIENT_REDIRECT_URI=http://localhost:3000/api/auth/callback/fims
-OIDC_MOCK_FORCED_ENABLE=true
 
 NEXT_PUBLIC_IS_MSW_ENABLED=true
